chore(refactor): refactor code
