+++ formily/next/playground/compiler/index.js
@@ -1,14 +1,11 @@
-// IMPORTS
-import { parseFormilyJSON } from './formilyParser'
+import { parse } from './parser'
 import { Flowise } from './flowise'
 
-function formily_flowise_compiler(formilyJSON) {
+function formilyCompiler(formilyJSON) {
   // Parsing formilyJSON -> Array of {title, component, description, validation(optional), options(optional)}
-  const fieldDetails = parseFormilyJSON(formilyJSON.schema.properties)
-
+  const fieldDetails = parse(formilyJSON.schema.properties)
   const flowiseJSON = new Flowise(fieldDetails)
-
   return flowiseJSON.toJSON()
 }
 
-export default formily_flowise_compiler
+export default formilyCompiler
