+++ formily/next/playground/compiler/logic/README.md
@@ -1,3 +1,10 @@
 # Logic
 
-This folder contains the internal logic of code runner transformers used in generating different form fields.
+This folder contains the code snippets for the Code Runner blocks being used for the recipe.
+
+## Index
+
+| File                             | Description                            |
+| -------------------------------- | -------------------------------------- |
+| [snippets.js](./snippets.js)     | Code for different code runner blocks  |
+| [validators.js](./validators.js) | Code for different validator functions |
