--- formily/next/playground/compiler/logic/generic.js
@@ -1,6 +1,0 @@
-/**
- * Generic code runner snippets
- */
-
-const INIT_CODE_RUNNER = `const msg = JSON.parse($0);`
-const EXIT_CODE_RUNNER = `return JSON.stringify(msg);`
