+++ formily/next/playground/widgets/ActionsWidget.tsx
@@ -11,15 +11,15 @@ export const ActionsWidget = observer(() => {
   useEffect(() => {
     loadInitialSchema(designer)
   }, [])
-  const supportLocales = ['zh-cn', 'en-us', 'ko-kr']
+  const supportLocales = ['en-us']
   useEffect(() => {
     if (!supportLocales.includes(GlobalRegistry.getDesignerLanguage())) {
-      GlobalRegistry.setDesignerLanguage('zh-cn')
+      GlobalRegistry.setDesignerLanguage('en-us')
     }
   }, [])
   return (
     <Space style={{ marginRight: 10 }}>
-      <Button href="https://designable-antd.formilyjs.org">Ant Design</Button>
+      {/* <Button href="https://designable-antd.formilyjs.org">Ant Design</Button> */}
       <Radio.Group
         value={GlobalRegistry.getDesignerLanguage()}
         optionType="button"
